abstracted table traits out to the EnhancedTable
